requirements.txt
streamlit
face_recognition
numpy
pillow
matplotlib